fixup alloca: fixed type issue

+++ include/bfdev/alloca.h
@@ -12,7 +12,7 @@
 BFDEV_BEGIN_DECLS
 
 #define bfdev_alloca __builtin_alloca
-extern __bfdev_malloc void *bfdev_alloca(unsigned int size);
+extern __bfdev_malloc void *bfdev_alloca(unsigned long size);
 
 BFDEV_END_DECLS
